# Improve FFI performance