--- a/cabal.project
+++ b/cabal.project
@@ -25,6 +25,20 @@ source-repository-package
     --sha256: 0sdw3ga7gbffyywpyx7xn8ghdw9gzsy4l68cas41hwfbkrwxwdg6
     tag: 32d7abec6a21c42a5f960d7f4133d604e8be79ec
 
+source-repository-package
+    type: git
+    location: https://github.com/hasktorch/inline-c
+    --sha256: 0hi4s3p0whf0ndsxn8qrfc3a7qndsgac0vz7n23kp3rx69yp0yr1
+    tag: 5b1a54e81f4c726faf35c15a8e39a4d86c52c6ba
+    subdir: inline-c
+
+source-repository-package
+    type: git
+    location: https://github.com/hasktorch/inline-c
+    --sha256: 0hi4s3p0whf0ndsxn8qrfc3a7qndsgac0vz7n23kp3rx69yp0yr1
+    tag: 5b1a54e81f4c726faf35c15a8e39a4d86c52c6ba
+    subdir: inline-c-cpp
+
 write-ghc-environment-files: always
 tests: true
 documentation: false