--- a/libtorch-ffi/libtorch-ffi.cabal
+++ b/libtorch-ffi/libtorch-ffi.cabal
@@ -138,7 +138,7 @@ library
  hs-source-dirs: src
  default-language: Haskell2010
  build-depends:       base >= 4.7 && < 5
-                    , inline-c-cpp >= 0.4.0.0 && < 0.5.0.0
+                    , inline-c-cpp >= 0.5.0.0 && < 0.6.0.0
                     , inline-c >= 0.9.0.0
                     , optparse-applicative >= 0.14.3.0
                     , containers