--- a/libtorch-ffi/src/Torch/Internal/Const.hs
+++ b/libtorch-ffi/src/Torch/Internal/Const.hs
@@ -11,7 +11,7 @@
 module Torch.Internal.Const where
 
 import qualified Language.C.Inline.Cpp as C
-import qualified Language.C.Inline.Cpp.Exceptions as C
+import qualified Language.C.Inline.Cpp.Unsafe as C
 import qualified Language.C.Inline.Context as C
 import qualified Language.C.Types as C
 import qualified Data.Map as Map