--- a/libtorch-ffi/src/Torch/Internal/Managed/Autograd.hs
+++ b/libtorch-ffi/src/Torch/Internal/Managed/Autograd.hs
@@ -12,11 +12,11 @@ import Foreign.C.Types (CBool)
 
 
 grad :: ForeignPtr Tensor -> ForeignPtr TensorList -> IO (ForeignPtr TensorList)
-grad = cast2 Unmanaged.grad
+grad = _cast2 Unmanaged.grad
 
 
 makeIndependent :: ForeignPtr Tensor -> CBool -> IO (ForeignPtr Tensor)
-makeIndependent = cast2 Unmanaged.makeIndependent
+makeIndependent = _cast2 Unmanaged.makeIndependent
 
 dropVariable :: ForeignPtr Tensor -> IO (ForeignPtr Tensor)
-dropVariable = cast1 Unmanaged.dropVariable
+dropVariable = _cast1 Unmanaged.dropVariable