--- a/libtorch-ffi/src/Torch/Internal/Managed/Type/C10Tuple.hs
+++ b/libtorch-ffi/src/Torch/Internal/Managed/Type/C10Tuple.hs
@@ -25,25 +25,25 @@ import qualified Torch.Internal.Unmanaged.Type.C10Tuple as Unmanaged
 
 newC10Tuple
   :: IO (ForeignPtr (C10Ptr IVTuple))
-newC10Tuple = cast0 Unmanaged.newC10Tuple
+newC10Tuple = _cast0 Unmanaged.newC10Tuple
 
 newC10Tuple_tuple
   :: ForeignPtr IValueList
   -> IO (ForeignPtr (C10Ptr IVTuple))
-newC10Tuple_tuple  = cast1 Unmanaged.newC10Tuple_tuple
+newC10Tuple_tuple  = _cast1 Unmanaged.newC10Tuple_tuple
 
 c10Tuple_empty
   :: ForeignPtr (C10Ptr IVTuple)
   -> IO (CBool)
-c10Tuple_empty = cast1 Unmanaged.c10Tuple_empty
+c10Tuple_empty = _cast1 Unmanaged.c10Tuple_empty
 
 c10Tuple_size
   :: ForeignPtr (C10Ptr IVTuple)
   -> IO (CSize)
-c10Tuple_size = cast1 Unmanaged.c10Tuple_size
+c10Tuple_size = _cast1 Unmanaged.c10Tuple_size
 
 c10Tuple_at
   :: ForeignPtr (C10Ptr IVTuple)
   -> CSize
   -> IO (ForeignPtr IValue)
-c10Tuple_at = cast2 Unmanaged.c10Tuple_at
+c10Tuple_at = _cast2 Unmanaged.c10Tuple_at