--- a/libtorch-ffi/src/Torch/Internal/Managed/Type/DimnameList.hs
+++ b/libtorch-ffi/src/Torch/Internal/Managed/Type/DimnameList.hs
@@ -26,27 +26,27 @@ import qualified Torch.Internal.Unmanaged.Type.DimnameList as Unmanaged
 
 newDimnameList
   :: IO (ForeignPtr DimnameList)
-newDimnameList = cast0 Unmanaged.newDimnameList
+newDimnameList = _cast0 Unmanaged.newDimnameList
 
 dimnameList_empty
   :: ForeignPtr DimnameList
   -> IO (CBool)
-dimnameList_empty = cast1 Unmanaged.dimnameList_empty
+dimnameList_empty = _cast1 Unmanaged.dimnameList_empty
 
 dimnameList_size
   :: ForeignPtr DimnameList
   -> IO (CSize)
-dimnameList_size = cast1 Unmanaged.dimnameList_size
+dimnameList_size = _cast1 Unmanaged.dimnameList_size
 
 dimnameList_at_s
   :: ForeignPtr DimnameList
   -> CSize
   -> IO (ForeignPtr Dimname)
-dimnameList_at_s = cast2 Unmanaged.dimnameList_at_s
+dimnameList_at_s = _cast2 Unmanaged.dimnameList_at_s
 
 dimnameList_push_back_n
   :: ForeignPtr DimnameList
   -> ForeignPtr Dimname
   -> IO (())
-dimnameList_push_back_n = cast2 Unmanaged.dimnameList_push_back_n
+dimnameList_push_back_n = _cast2 Unmanaged.dimnameList_push_back_n