--- a/libtorch-ffi/src/Torch/Internal/Managed/Type/IValueList.hs
+++ b/libtorch-ffi/src/Torch/Internal/Managed/Type/IValueList.hs
@@ -25,7 +25,7 @@ import qualified Torch.Internal.Unmanaged.Type.IValueList as Unmanaged
 
 newIValueList
   :: IO (ForeignPtr IValueList)
-newIValueList = cast0 Unmanaged.newIValueList
+newIValueList = _cast0 Unmanaged.newIValueList
 
 
 
@@ -34,24 +34,24 @@ newIValueList = cast0 Unmanaged.newIValueList
 ivalueList_empty
   :: ForeignPtr IValueList
   -> IO (CBool)
-ivalueList_empty = cast1 Unmanaged.ivalueList_empty
+ivalueList_empty = _cast1 Unmanaged.ivalueList_empty
 
 ivalueList_size
   :: ForeignPtr IValueList
   -> IO (CSize)
-ivalueList_size = cast1 Unmanaged.ivalueList_size
+ivalueList_size = _cast1 Unmanaged.ivalueList_size
 
 ivalueList_at
   :: ForeignPtr IValueList
   -> CSize
   -> IO (ForeignPtr IValue)
-ivalueList_at = cast2 Unmanaged.ivalueList_at
+ivalueList_at = _cast2 Unmanaged.ivalueList_at
 
 ivalueList_push_back
   :: ForeignPtr IValueList
   -> ForeignPtr IValue
   -> IO (())
-ivalueList_push_back = cast2 Unmanaged.ivalueList_push_back
+ivalueList_push_back = _cast2 Unmanaged.ivalueList_push_back