--- a/libtorch-ffi/src/Torch/Internal/Managed/Type/IntArray.hs
+++ b/libtorch-ffi/src/Torch/Internal/Managed/Type/IntArray.hs
@@ -26,27 +26,27 @@ import qualified Torch.Internal.Unmanaged.Type.IntArray as Unmanaged
 
 newIntArray
   :: IO (ForeignPtr IntArray)
-newIntArray = cast0 Unmanaged.newIntArray
+newIntArray = _cast0 Unmanaged.newIntArray
 
 intArray_empty
   :: ForeignPtr IntArray
   -> IO (CBool)
-intArray_empty = cast1 Unmanaged.intArray_empty
+intArray_empty = _cast1 Unmanaged.intArray_empty
 
 intArray_size
   :: ForeignPtr IntArray
   -> IO (CSize)
-intArray_size = cast1 Unmanaged.intArray_size
+intArray_size = _cast1 Unmanaged.intArray_size
 
 intArray_at_s
   :: ForeignPtr IntArray
   -> CSize
   -> IO (Int64)
-intArray_at_s = cast2 Unmanaged.intArray_at_s
+intArray_at_s = _cast2 Unmanaged.intArray_at_s
 
 intArray_push_back_l
   :: ForeignPtr IntArray
   -> Int64
   -> IO (())
-intArray_push_back_l = cast2 Unmanaged.intArray_push_back_l
+intArray_push_back_l = _cast2 Unmanaged.intArray_push_back_l