Calling FFI is slow. It takes more than 1us.
For example, even a small tensor takes more than 1us.

Here are the details inside.

|  Operation  |  Operating Time  |　Link |
| ---- | ---- | ---- |
|  libtorch's adder  |  900ns | https://discuss.pytorch.org/t/too-much-c-api-overhead/164381 |
|  newForeignPtr with cfinalizer  |  80ns | https://gitlab.haskell.org/ghc/ghc/-/issues/22394 |
|  Safe call of FFI |  100ns | https://qiita.com/mod_poppo/items/793fdb08e62591d6f3fb  |
|  Storing Exception of C++ |  100ns | https://github.com/fpco/inline-c/pull/140  |
|  Useless cast  where no exception occurs|  100ns  |   |

This PR improves as follows.

|  Operation  |  Operating Time  |　Link |
| ---- | ---- | ---- |
|  Safe call of FFI |  100ns  -> 10ns | https://qiita.com/mod_poppo/items/793fdb08e62591d6f3fb  |
|  Storing Exception of C++ |  100ns  -> 30ns |   https://github.com/fpco/inline-c/pull/140 |
|  Useless cast  where no exception occurs|  100ns -> 0ns  |   |