@junjihashimoto 2022-11-11
# Hasktorch vs Hmatrix

## Before

![image](https://user-images.githubusercontent.com/2469618/201331473-5d917bd2-f7f4-4e22-93cf-220a3cc898aa.png)

## After this PR

![image](https://user-images.githubusercontent.com/2469618/201331235-3481615a-36db-4a69-91e6-e921a5746819.png)